renamed models file to usermodel
